feat: Gcloud exec with connection support

+++ checks/exec.go
@@ -2,6 +2,7 @@ package checks
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"os"
 	osExec "os/exec"
@@ -70,10 +71,14 @@ func execBash(check v1.ExecCheck, ctx *context.Context) pkg.Results {
 	// Setup connection details
 	switch strings.ToLower(fields[0]) {
 	case "aws":
+		if check.Connections.AWS == nil {
+			return []*pkg.CheckResult{result.Failf("no AWS connection provided")}
+		}
+
 		configPath, err := createAWSCredentialsFile(ctx, check.Connections.AWS)
 		defer os.RemoveAll(configPath)
 		if err != nil {
-			return []*pkg.CheckResult{result.Failf(err.Error())}
+			return []*pkg.CheckResult{result.Failf("failed to store AWS credentials: %v", err.Error())}
 		}
 
 		cmd.Env = os.Environ()
@@ -82,6 +87,25 @@ func execBash(check v1.ExecCheck, ctx *context.Context) pkg.Results {
 	case "az":
 
 	case "gcloud":
+		if check.Connections.GCP == nil {
+			return []*pkg.CheckResult{result.Failf("no GCP connection provided")}
+		}
+
+		configPath, err := createGCloudCredentialsFile(ctx, check.Connections.GCP)
+		defer os.RemoveAll(configPath)
+		if err != nil {
+			return []*pkg.CheckResult{result.Failf("failed to store gcloud credentials: %v", err.Error())}
+		}
+
+		// to configure gcloud CLI to use the service account specified in GOOGLE_APPLICATION_CREDENTIALS,
+		// we need to explicitly activate it
+		runCmd := osExec.Command("gcloud", "auth", "activate-service-account", "--key-file", configPath)
+		if err := runCmd.Run(); err != nil {
+			return []*pkg.CheckResult{result.Failf("failed to activate GCP service account: %v", err.Error())}
+		}
+
+		cmd.Env = os.Environ()
+		cmd.Env = append(cmd.Env, fmt.Sprintf("GOOGLE_APPLICATION_CREDENTIALS=%s", configPath))
 	}
 
 	return runCmd(cmd, result)
@@ -106,18 +130,34 @@ func runCmd(cmd *osExec.Cmd, result *pkg.CheckResult) (results pkg.Results) {
 	return results
 }
 
+func createGCloudCredentialsFile(ctx *context.Context, conn *v1.GCPConnection) (string, error) {
+	if err := conn.HydrateConnection(ctx); err != nil {
+		return "", err
+	}
+
+	if conn.Credentials.ValueStatic == "" {
+		return "", errors.New("gcp credentials is empty")
+	}
+
+	return saveConfig("gcloud-*", gcloudConfigTemplate, conn)
+}
+
 func createAWSCredentialsFile(ctx *context.Context, conn *v1.AWSConnection) (string, error) {
 	if err := conn.Populate(ctx, ctx.Kubernetes, ctx.Namespace); err != nil {
 		return "", err
 	}
 
-	dirPath, err := os.MkdirTemp("", "aws-*")
+	return saveConfig("aws-*", awsConfigTemplate, conn)
+}
+
+func saveConfig(dirPrefix string, configTemplate *textTemplate.Template, view any) (string, error) {
+	dirPath, err := os.MkdirTemp("", dirPrefix)
 	if err != nil {
 		return "", err
 	}
 
 	configPath := fmt.Sprintf("%s/credentials", dirPath)
-	logger.Tracef("Creating AWS credentials file: %s", configPath)
+	logger.Tracef("Creating credentials file: %s", configPath)
 
 	file, err := os.Create(configPath)
 	if err != nil {
@@ -125,7 +165,7 @@ func createAWSCredentialsFile(ctx *context.Context, conn *v1.AWSConnection) (str
 	}
 	defer file.Close()
 
-	if err := awsConfigTemplate.Execute(file, conn); err != nil {
+	if err := configTemplate.Execute(file, view); err != nil {
 		return configPath, err
 	}
 
@@ -133,16 +173,15 @@ func createAWSCredentialsFile(ctx *context.Context, conn *v1.AWSConnection) (str
 }
 
 var (
-	awsConfigTemplate *textTemplate.Template
+	awsConfigTemplate    *textTemplate.Template
+	gcloudConfigTemplate *textTemplate.Template
 )
 
 func init() {
-	const (
-		awsConfigTpl = `[default]
+	awsConfigTemplate = textTemplate.Must(textTemplate.New("").Parse(`[default]
 aws_access_key_id = {{.AccessKey}}
 aws_secret_access_key = {{.SecretKey}}
-{{if .Region}}region = {{.Region}}{{end}}`
-	)
+{{if .Region}}region = {{.Region}}{{end}}`))
 
-	awsConfigTemplate = textTemplate.Must(textTemplate.New("").Parse(awsConfigTpl))
+	gcloudConfigTemplate = textTemplate.Must(textTemplate.New("").Parse(`{{.Credentials}}`))
 }
